finance_backend/requirements.txt
Django==3.2.25
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
firebase-admin==6.4.0
google-cloud-firestore==2.13.1
python-dotenv==1.0.0
asyncio==3.4.3 